moved create_ap to be a plugin.... it was easier than even I thought it would be.
