+++ copilot/plugins/create_ap/plugin.conf
@@ -1,0 +1,4 @@
+[info]
+name = create_ap
+config_file = ap.conf
+directory = /tmp/copilot/
